feat: common attachment chips
