+++ app/javascript/dashboard/components-next/message/bubbles/Attachments.vue
@@ -1,13 +1,7 @@
 <script setup>
-import { computed } from 'vue';
-
 import BaseBubble from 'next/message/bubbles/Base.vue';
-import ImageChip from 'next/message/chips/Image.vue';
-import VideoChip from 'next/message/chips/Video.vue';
-import AudioChip from 'next/message/chips/Audio.vue';
-import FileChip from 'next/message/chips/File.vue';
+import AttachmentChips from 'next/message/chips/AttachmentChips.vue';
 
-import { ATTACHMENT_TYPES } from '../constants';
 /**
  * @typedef {Object} Attachment
  * @property {number} id - Unique identifier for the attachment
@@ -21,66 +15,16 @@ import { ATTACHMENT_TYPES } from '../constants';
  * @property {number|null} width - Width of the image if applicable
  * @property {number|null} height - Height of the image if applicable
  */
-const props = defineProps({
+defineProps({
   attachments: {
     type: Array,
     default: () => [],
   },
 });
-
-const mediaAttachments = computed(() => {
-  const allowedTypes = [ATTACHMENT_TYPES.IMAGE, ATTACHMENT_TYPES.VIDEO];
-  const mediaTypes = props.attachments.filter(attachment =>
-    allowedTypes.includes(attachment.fileType)
-  );
-
-  return mediaTypes.sort(
-    (a, b) =>
-      allowedTypes.indexOf(a.fileType) - allowedTypes.indexOf(b.fileType)
-  );
-});
-
-const recordings = computed(() => {
-  return props.attachments.filter(
-    attachment => attachment.fileType === ATTACHMENT_TYPES.AUDIO
-  );
-});
-
-const files = computed(() => {
-  return props.attachments.filter(
-    attachment => attachment.fileType === ATTACHMENT_TYPES.FILE
-  );
-});
 </script>
 
 <template>
   <BaseBubble class="grid gap-2 bg-transparent">
-    <div v-if="mediaAttachments.length" class="flex gap-1">
-      <template v-for="attachment in mediaAttachments" :key="attachment.id">
-        <ImageChip
-          v-if="attachment.fileType === ATTACHMENT_TYPES.IMAGE"
-          :attachment="attachment"
-        />
-        <VideoChip
-          v-else-if="attachment.fileType === ATTACHMENT_TYPES.VIDEO"
-          :attachment="attachment"
-        />
-      </template>
-    </div>
-    <div v-if="recordings.length" class="flex flex-wrap gap-1">
-      <AudioChip
-        v-for="attachment in recordings"
-        :key="attachment.id"
-        class="bg-n-alpha-3 dark:bg-n-alpha-2 text-n-slate-12"
-        :attachment="attachment"
-      />
-    </div>
-    <div v-if="files.length" class="flex flex-wrap gap-1">
-      <FileChip
-        v-for="attachment in files"
-        :key="attachment.id"
-        :attachment="attachment"
-      />
-    </div>
+    <AttachmentChips :attachments="attachments" class="gap-1" />
   </BaseBubble>
 </template>
