+++ app/javascript/dashboard/components-next/message/bubbles/Email.vue
@@ -3,14 +3,9 @@ import { computed, useTemplateRef, ref, watch } from 'vue';
 import { Letter } from 'vue-letter';
 import Icon from 'next/icon/Icon.vue';
 import BaseBubble from 'next/message/bubbles/Base.vue';
-
-import ImageChip from 'next/message/chips/Image.vue';
-import VideoChip from 'next/message/chips/Video.vue';
-import AudioChip from 'next/message/chips/Audio.vue';
-import FileChip from 'next/message/chips/File.vue';
+import AttachmentChips from 'next/message/chips/AttachmentChips.vue';
 
 import { MESSAGE_STATUS } from '../constants';
-import { ATTACHMENT_TYPES } from '../constants';
 
 const props = defineProps({
   content: {
@@ -101,30 +96,6 @@ const showMeta = computed(() => {
     subject.value
   );
 });
-
-const mediaAttachments = computed(() => {
-  const allowedTypes = [ATTACHMENT_TYPES.IMAGE, ATTACHMENT_TYPES.VIDEO];
-  const mediaTypes = props.attachments.filter(attachment =>
-    allowedTypes.includes(attachment.fileType)
-  );
-
-  return mediaTypes.sort(
-    (a, b) =>
-      allowedTypes.indexOf(a.fileType) - allowedTypes.indexOf(b.fileType)
-  );
-});
-
-const recordings = computed(() => {
-  return props.attachments.filter(
-    attachment => attachment.fileType === ATTACHMENT_TYPES.AUDIO
-  );
-});
-
-const files = computed(() => {
-  return props.attachments.filter(
-    attachment => attachment.fileType === ATTACHMENT_TYPES.FILE
-  );
-});
 </script>
 
 <template>
@@ -182,33 +153,7 @@ const files = computed(() => {
       />
     </section>
     <section class="px-4 pb-4">
-      <div v-if="mediaAttachments.length" class="flex flex-wrap gap-2">
-        <template v-for="attachment in mediaAttachments" :key="attachment.id">
-          <ImageChip
-            v-if="attachment.fileType === ATTACHMENT_TYPES.IMAGE"
-            :attachment="attachment"
-          />
-          <VideoChip
-            v-else-if="attachment.fileType === ATTACHMENT_TYPES.VIDEO"
-            :attachment="attachment"
-          />
-        </template>
-      </div>
-      <div v-if="recordings.length" class="flex flex-wrap gap-2">
-        <AudioChip
-          v-for="attachment in recordings"
-          :key="attachment.id"
-          class="bg-n-alpha-3 dark:bg-n-alpha-2 text-n-slate-12"
-          :attachment="attachment"
-        />
-      </div>
-      <div v-if="files.length" class="flex flex-wrap gap-2">
-        <FileChip
-          v-for="attachment in files"
-          :key="attachment.id"
-          :attachment="attachment"
-        />
-      </div>
+      <AttachmentChips :attachments="attachments" class="gap-1" />
     </section>
   </BaseBubble>
 </template>
